requirements.txt
fastapi
uvicorn
mysql-connector-python
transformers
torch
pydantic
python-dotenv
